Refactor: 사이드바 구조화 작업
