+++ client/src/components/layout/sidebar/hooks/index.ts
@@ -1,0 +1,4 @@
+import useGroupedChats from "./useGroupedChats";
+import useChatRooms from "./useChatRooms";
+
+export { useChatRooms, useGroupedChats };
