+++ client/src/components/layout/sidebar/models/index.ts
@@ -1,4 +1,5 @@
 import ChatRoom from "./ChatRoom";
 import ChatRoomMenu from "./ChatRoomMenu";
+import SidebarContent from "./SidebarContent";
 
-export { ChatRoom, ChatRoomMenu };
+export { ChatRoom, ChatRoomMenu, SidebarContent };
